+++ db/migrations/20201026204012-create-watchlist.js
@@ -17,7 +17,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       currentPrice: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC(10, 2)
       },
       userId: {
         allowNull: false,
